(api-cat) Rename rest controllers according to api paths
